修复该模板已关联渠道，暂不可删除

+++ hyjf_am_server/hyjf_am_core/hyjf_am_admin/src/main/java/com/hyjf/am/user/controller/admin/promotion/LandingManageController.java
@@ -111,11 +111,18 @@ public class LandingManageController extends BaseController {
     public TemplateConfigResponse deleteTemplate(@PathVariable Integer id) {
         logger.info("---deleteTemplate by param---  " + JSONObject.toJSON(id));
         TemplateConfigResponse response = new TemplateConfigResponse();
+        int count=landingManagerService.deleteTemplateCount(id);
+        if(count>0) {
+        	response.setRtn(UtmResponse.FAIL);
+        	response.setMessage("该模板已关联渠道，暂不可删除");
+        	return response;
+        }
         try {
             int intDelete = landingManagerService.deleteTemplate(id);
             response.setRtn(UtmResponse.SUCCESS);
         } catch (Exception e) {
             response.setRtn(UtmResponse.FAIL);
+            response.setMessage("删除失败");
         }
         return response;
     }
